导入功能开发。

+++ ruoyi-email/src/main/java/com/ruoyi/email/mapper/ImportEmailMapper.java
@@ -66,7 +66,7 @@ public interface ImportEmailMapper
      * @param createId
      * @return
      */
-    long count(@Param("createId") Long createId);
+    int count(@Param("createId") Long createId);
 
     /**
      * 导入邮件列表
